backend/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart
pydantic
requests
aiohttp
google-genai
Pillow
numpy
python-dotenv
gTTS
flask-cors
moviepy
imageio
imageio-ffmpeg
openai
setuptools